[-] MO : blockwishlist, many bugs fixed

+++ modules/blockwishlist/mywishlist.php
@@ -16,7 +16,7 @@ if ($cookie->isLogged())
 	$delete = Tools::getIsset('deleted');
 	$delete = (empty($delete) === false ? 1 : 0);
 	$id_wishlist = Tools::getValue('id_wishlist');
-	if (isset($_POST['submitWishlist']))
+	if (Tools::isSubmit('submitWishlist'))
 	{
 		if (Configuration::get('PS_TOKEN_ACTIVATED') == 1 AND
 			strcmp(Tools::getToken(), Tools::getValue('token')))
@@ -24,13 +24,21 @@ if ($cookie->isLogged())
 		if (!sizeof($errors))
 		{
 			$name = Tools::getValue('name');
-			$wishlist = new WishList();
-			$wishlist->name = $name;
-			$wishlist->id_customer = $cookie->id_customer;
-			list($us, $s) = explode(' ', microtime());
-			srand($s * $us);
-			$wishlist->token = strtoupper(substr(sha1(uniqid(rand(), true)._COOKIE_KEY_.$cookie->id_customer), 0, 16));
-			$wishlist->add();
+			if (empty($name))
+				$errors[] = Tools::displayError('you must specify a name');
+			if (WishList::isExistsByNameForUser($name))
+				$errors[] = Tools::displayError('this name is already used by an other list');
+			
+			if(!sizeof($errors))
+			{
+				$wishlist = new WishList();
+				$wishlist->name = $name;
+				$wishlist->id_customer = $cookie->id_customer;
+				list($us, $s) = explode(' ', microtime());
+				srand($s * $us);
+				$wishlist->token = strtoupper(substr(sha1(uniqid(rand(), true)._COOKIE_KEY_.$cookie->id_customer), 0, 16));
+				$wishlist->add();
+			}
 		}
 	}
 	else if ($add)
@@ -38,14 +46,17 @@ if ($cookie->isLogged())
 	else if ($delete AND empty($id_wishlist) === false)
 	{
 		$wishlist = new WishList(intval($id_wishlist));
-		$wishlist->delete();
+		if (Validate::isLoadedObject($wishlist))
+			$wishlist->delete();
+		else
+			$errors[] = Tools::displayError('can`t delete this whislist');
 	}
 	$smarty->assign('wishlists', WishList::getByIdCustomer(intval($cookie->id_customer)));
 	$smarty->assign('nbProducts', WishList::getInfosByIdCustomer(intval($cookie->id_customer)));
 }
 else
 {
-	$errors[] = Tools::displayError('You need to be logged to manage your wishlist'); 
+	Tools::redirect('authentication.php?back=/modules/blockwishlist/mywishlist.php');
 }
 
 $smarty->assign('id_customer', intval($cookie->id_customer));
